refactor(performance): fold circular-dependency hints into one SCC scan
