+++ lib/CircularModulesPlugin.js
@@ -5,13 +5,21 @@
 
 "use strict";
 
+const CircularDependenciesWarning = require("./errors/CircularDependenciesWarning");
+const { compareModulesByIdentifier } = require("./util/comparators");
+
+/** @typedef {import("../declarations/WebpackOptions").PerformanceOptions} PerformanceOptions */
 /** @typedef {import("./Compiler")} Compiler */
 /** @typedef {import("./ModuleGraph")} ModuleGraph */
 /** @typedef {import("./ModuleGraphConnection")} ModuleGraphConnection */
 /** @typedef {import("./Module")} Module */
 /** @typedef {import("./Module").BuildInfo} BuildInfo */
+/** @typedef {import("./RequestShortener")} RequestShortener */
 
 const PLUGIN_NAME = "CircularModulesPlugin";
+
+// Enough to name the worst tangles without printing the module graph.
+const MAX_REPORTED_CYCLES = 5;
 
 /**
  * The module a connection makes its origin evaluate before it can run, or
@@ -40,6 +48,80 @@ const getSynchronousTarget = (connection, module, moduleGraph) => {
 };
 
 /**
+ * The member a group is reported from. Traversal order is not stable across
+ * runs, so the lowest identifier is picked rather than the one found first.
+ * @param {Module[]} members one group of modules that can all reach each other
+ * @returns {Module} the module the report starts at
+ */
+const getCanonicalMember = (members) => {
+	let canonical = members[0];
+
+	for (const member of members) {
+		if (compareModulesByIdentifier(member, canonical) < 0) {
+			canonical = member;
+		}
+	}
+
+	return canonical;
+};
+
+/**
+ * The shortest cycle through the canonical module of a group, as a readable
+ * path ending back where it started.
+ * @param {Module[]} members one group of modules that can all reach each other
+ * @param {ModuleGraph} moduleGraph the module graph
+ * @param {RequestShortener} requestShortener the request shortener
+ * @returns {string} the cycle path
+ */
+const formatShortestCycle = (members, moduleGraph, requestShortener) => {
+	const group = new Set(members);
+	const start = getCanonicalMember(members);
+	/** @type {Map<Module, Module>} */
+	const previous = new Map();
+	const queue = [start];
+	/** @type {Module | undefined} */
+	let last;
+
+	for (let i = 0; i < queue.length && last === undefined; i++) {
+		const module = queue[i];
+
+		for (const connection of moduleGraph.getOutgoingConnections(module)) {
+			const target = getSynchronousTarget(connection, module, moduleGraph);
+
+			// A module reading its own exports would otherwise collapse the path.
+			if (target === null || target === module || !group.has(target)) continue;
+
+			if (target === start) {
+				last = module;
+				break;
+			}
+
+			if (previous.has(target)) continue;
+
+			previous.set(target, module);
+			queue.push(target);
+		}
+	}
+
+	const path = [start];
+
+	for (
+		let module = last;
+		module !== undefined && module !== start;
+		module = previous.get(module)
+	) {
+		path.push(module);
+	}
+
+	path.push(start);
+
+	return path
+		.reverse()
+		.map((module) => module.readableIdentifier(requestShortener))
+		.join(" -> ");
+};
+
+/**
  * Detects circular dependencies among synchronous module imports.
  *
  * Builds an adjacency layout from each module's synchronous outgoing
@@ -52,14 +134,14 @@ const getSynchronousTarget = (connection, module, moduleGraph) => {
  */
 class CycleGraph {
 	/**
-	 * @param {Set<Module>} circularModules modules involved in circular dependencies
-	 * @param {Module[][]} components groups of at least two modules that can all reach each other
+	 * @param {Set<Module>} circularModules modules in a multi-module SCC or with a self-loop (for isCircular)
+	 * @param {Module[][]} circularGroups multi-module SCC groups only (for reporting; self-loops omitted)
 	 */
-	constructor(circularModules, components) {
+	constructor(circularModules, circularGroups) {
 		/** @type {Set<Module>} */
 		this.circularModules = circularModules;
 		/** @type {Module[][]} */
-		this.components = components;
+		this.circularGroups = circularGroups;
 	}
 
 	/**
@@ -108,9 +190,9 @@ class CycleGraph {
 
 		// Iterative SCC algorithm
 		/** @type {Set<Module>} */
-		const circular = new Set();
+		const circularModules = new Set();
 		/** @type {Module[][]} */
-		const components = [];
+		const circularGroups = [];
 		let nextIndex = 0;
 		const nodeIndex = new Int32Array(size).fill(-1);
 		const nodeLowLink = new Int32Array(size);
@@ -157,24 +239,24 @@ class CycleGraph {
 				} else {
 					if (nodeLowLink[v] === nodeIndex[v]) {
 						/** @type {number[]} */
-						const component = [];
+						const group = [];
 						let w;
 						do {
 							w = /** @type {number} */ (sccStack.pop());
 							nodeOnStack[w] = 0;
-							component.push(w);
+							group.push(w);
 						} while (w !== v);
 
-						if (component.length > 1 || selfLoops[v]) {
-							for (const idx of component) {
-								circular.add(moduleList[idx]);
+						if (group.length > 1 || selfLoops[v]) {
+							for (const idx of group) {
+								circularModules.add(moduleList[idx]);
 							}
 						}
 
 						// A module that only references itself is no cycle, so it is
 						// circular for inlining but never a group to report.
-						if (component.length > 1) {
-							components.push(component.map((idx) => moduleList[idx]));
+						if (group.length > 1) {
+							circularGroups.push(group.map((idx) => moduleList[idx]));
 						}
 					}
 
@@ -189,29 +271,82 @@ class CycleGraph {
 			}
 		}
 
-		return new CycleGraph(circular, components);
+		return new CycleGraph(circularModules, circularGroups);
 	}
 }
 
 /**
- * Detects circular dependencies and marks each circular module
- * via buildInfo.isCircular for downstream consumers.
+ * @typedef {object} CircularModulesPluginOptions
+ * @property {PerformanceOptions["hints"]=} hints when set, report SCC groups
+ */
+
+/**
+ * One SCC scan: marks isCircular on every module, and reports groups when hints are set.
  */
 class CircularModulesPlugin {
+	/**
+	 * @param {CircularModulesPluginOptions=} options options
+	 */
+	constructor(options) {
+		/** @type {PerformanceOptions["hints"] | undefined} */
+		this.hints = options && options.hints;
+	}
+
 	/**
 	 * @param {Compiler} compiler the compiler instance
 	 * @returns {void}
 	 */
 	apply(compiler) {
+		const hints = this.hints;
+
 		compiler.hooks.compilation.tap(PLUGIN_NAME, (compilation) => {
 			compilation.hooks.optimizeModules.tap(PLUGIN_NAME, (modules) => {
-				const { circularModules } = CycleGraph.build(
+				const { circularModules, circularGroups } = CycleGraph.build(
 					modules,
 					compilation.moduleGraph
 				);
+
+				// Must be an explicit boolean: ConstExportsPlugin checks `=== false`.
 				for (const m of modules) {
 					/** @type {BuildInfo} */
 					(m.buildInfo).isCircular = circularModules.has(m);
+				}
+
+				if (!hints) return;
+				if (circularGroups.length === 0) return;
+
+				const { moduleGraph, requestShortener } = compilation;
+
+				// Largest tangle first; ties break by name, as the order groups are
+				// discovered in is not stable.
+				circularGroups.sort(
+					(a, b) =>
+						b.length - a.length ||
+						compareModulesByIdentifier(
+							getCanonicalMember(a),
+							getCanonicalMember(b)
+						)
+				);
+
+				// The shortest cycle of a large group names two of its modules, so the
+				// size travels with it — otherwise the group reads as a pair.
+				const groups = circularGroups
+					.slice(0, MAX_REPORTED_CYCLES)
+					.map((members) => ({
+						size: members.length,
+						cycle: formatShortestCycle(members, moduleGraph, requestShortener)
+					}));
+				const warning = new CircularDependenciesWarning(
+					groups,
+					circularGroups.length
+				);
+
+				if (hints === "error") {
+					compilation.errors.push(warning);
+				} else if (hints === "stats") {
+					compilation.hints.push(warning);
+				} else {
+					compilation.warnings.push(warning);
 				}
 			});
 		});
@@ -219,5 +354,3 @@ class CircularModulesPlugin {
 }
 
 module.exports = CircularModulesPlugin;
-module.exports.CycleGraph = CycleGraph;
-module.exports.getSynchronousTarget = getSynchronousTarget;
