+++ lib/WebpackOptionsApply.js
@@ -724,10 +724,15 @@ class WebpackOptionsApply extends OptionsApply {
 			new InnerGraphPlugin().apply(compiler);
 		}
 
-		if (options.mode === "production") {
+		const circularHints =
+			options.performance && options.performance.circularDependencies
+				? options.performance.hints
+				: false;
+
+		if (options.mode === "production" || circularHints) {
 			const CircularModulesPlugin = require("./CircularModulesPlugin");
 
-			new CircularModulesPlugin().apply(compiler);
+			new CircularModulesPlugin({ hints: circularHints }).apply(compiler);
 		}
 
 		const ConstExportsPlugin = require("./optimize/ConstExportsPlugin");
@@ -927,12 +932,6 @@ class WebpackOptionsApply extends OptionsApply {
 				const CacheEffectivenessPlugin = require("./performance/CacheEffectivenessPlugin");
 
 				new CacheEffectivenessPlugin(options.performance).apply(compiler);
-			}
-
-			if (options.performance.circularDependencies) {
-				const CircularDependenciesPlugin = require("./performance/CircularDependenciesPlugin");
-
-				new CircularDependenciesPlugin(options.performance).apply(compiler);
 			}
 
 			if (options.performance.duplicateModules) {
